feat: dbaas - implemented the parameters resource
